# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Cosmic Web Visualization API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Galaxy {
  /** Cartesian X position in Mpc (Earth at origin) */
  x: number;
  /** Cartesian Y position in Mpc */
  y: number;
  /** Cartesian Z position in Mpc */
  z: number;
  /** Right ascension in degrees */
  ra: number;
  /** Declination in degrees */
  dec: number;
  /** Spectroscopic redshift */
  redshift: number;
  /** Comoving distance in Mpc */
  distance: number;
  /** Local density contrast (delta = rho/rho_mean - 1) */
  density: number;
  /** Source dataset label */
  dataset: string;
}

export interface Cosmology {
  /** Hubble constant km/s/Mpc */
  H0: number;
  /** Matter density parameter */
  OmegaM: number;
  /** Dark energy density parameter */
  OmegaL: number;
}

export interface GalaxyResponse {
  galaxies: Galaxy[];
  /** Total galaxies in filtered set */
  total: number;
  /** Number of galaxies returned */
  returned: number;
  cosmology: Cosmology;
}

export type GalaxyStatsZRange = {
  min: number;
  max: number;
};

export interface DatasetInfo {
  name: string;
  count: number;
  zMin: number;
  zMax: number;
  label: string;
}

export interface GalaxyStats {
  datasets: DatasetInfo[];
  totalGalaxies: number;
  zRange: GalaxyStatsZRange;
}

export interface DensityCell {
  x: number;
  y: number;
  z: number;
  density: number;
  vx: number;
  vy: number;
  vz: number;
}

export type DensityGridBounds = {
  min: number;
  max: number;
};

export interface DensityGrid {
  cells: DensityCell[];
  resolution: number;
  bounds: DensityGridBounds;
}

export type GetGalaxiesParams = {
/**
 * Max number of galaxies to return
 */
sample?: number;
/**
 * Minimum redshift
 */
zmin?: number;
/**
 * Maximum redshift
 */
zmax?: number;
/**
 * Dataset filter (all, lrg60, lrg70, lrg80, lrg90, lrg100, lrg2)
 */
dataset?: string;
};

export type GetDensityGridParams = {
/**
 * Grid resolution per axis
 */
resolution?: number;
};