# lib/api-zod/src/generated/types/cosmology.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Cosmic Web Visualization API
 * OpenAPI spec version: 0.1.0
 */

export interface Cosmology {
  /** Hubble constant km/s/Mpc */
  H0: number;
  /** Matter density parameter */
  OmegaM: number;
  /** Dark energy density parameter */
  OmegaL: number;
}