# lib/api-zod/src/generated/types/datasetInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Cosmic Web Visualization API
 * OpenAPI spec version: 0.1.0
 */

export interface DatasetInfo {
  name: string;
  count: number;
  zMin: number;
  zMax: number;
  label: string;
}