# lib/api-zod/src/generated/types/densityCell.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Cosmic Web Visualization API
 * OpenAPI spec version: 0.1.0
 */

export interface DensityCell {
  x: number;
  y: number;
  z: number;
  density: number;
  vx: number;
  vy: number;
  vz: number;
}