# lib/api-zod/src/generated/types/densityGrid.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Cosmic Web Visualization API
 * OpenAPI spec version: 0.1.0
 */
import type { DensityCell } from './densityCell';
import type { DensityGridBounds } from './densityGridBounds';

export interface DensityGrid {
  cells: DensityCell[];
  resolution: number;
  bounds: DensityGridBounds;
}