# lib/api-zod/src/generated/types/galaxy.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Cosmic Web Visualization API
 * OpenAPI spec version: 0.1.0
 */

export interface Galaxy {
  /** Cartesian X position in Mpc (Earth at origin) */
  x: number;
  /** Cartesian Y position in Mpc */
  y: number;
  /** Cartesian Z position in Mpc */
  z: number;
  /** Right ascension in degrees */
  ra: number;
  /** Declination in degrees */
  dec: number;
  /** Spectroscopic redshift */
  redshift: number;
  /** Comoving distance in Mpc */
  distance: number;
  /** Local density contrast (delta = rho/rho_mean - 1) */
  density: number;
  /** Source dataset label */
  dataset: string;
}