# lib/api-zod/src/generated/types/galaxyResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Cosmic Web Visualization API
 * OpenAPI spec version: 0.1.0
 */
import type { Cosmology } from './cosmology';
import type { Galaxy } from './galaxy';

export interface GalaxyResponse {
  galaxies: Galaxy[];
  /** Total galaxies in filtered set */
  total: number;
  /** Number of galaxies returned */
  returned: number;
  cosmology: Cosmology;
}