# lib/api-zod/src/generated/types/galaxyStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Cosmic Web Visualization API
 * OpenAPI spec version: 0.1.0
 */
import type { DatasetInfo } from './datasetInfo';
import type { GalaxyStatsZRange } from './galaxyStatsZRange';

export interface GalaxyStats {
  datasets: DatasetInfo[];
  totalGalaxies: number;
  zRange: GalaxyStatsZRange;
}